First cut of rewrite of very messy session management. Removal of RTP "encoding" sbusystem and the separate RTP_DataObject attached to RTP sessions, so now have a simple single class hierarchy
git-svn-id: svn://svn.code.sf.net/p/opalvoip/code/opal/branches/Luyten_A@24786 023b2edf-31b2-4de3-b41e-bca80c47788f

+++ samples/simple/main.cxx
@@ -628,7 +628,7 @@ PBoolean MyManager::Initialise(PArgList & args)
     cerr << " port is " << port << endl;
 
     if (port > 0)
-      iax2EP = new IAX2EndPoint(*this, port);
+      iax2EP = new IAX2EndPoint(*this, (WORD)port);
     else
       iax2EP = new IAX2EndPoint(*this);
